Refactor so shard / non-shard cases can use the same code
The shard / non-shard cases in CRA use different API. Ideally,
we would have a base class for the common API and have the
shard case inherit from the base class + either VertexBase or
ShardedVertexBase. However, C# does not support multiple
inheritence.

To get around this, we move the initialized methods into a sharded
and non-sharded case. Both of these cases call the original
Ambrosia initialize function.

To make this cleaner and potentially easier to test, we move this
logic into a different file.
