+++ Ambrosia/Ambrosia/Program.cs
@@ -1042,29 +1042,7 @@ namespace Ambrosia
         }
     }
 
-    public class AmbrosiaRuntimeParams
-    {
-        public int serviceReceiveFromPort;
-        public int serviceSendToPort;
-        public string serviceName;
-        public string AmbrosiaBinariesLocation;
-        public string serviceLogPath;
-        public bool? createService;
-        public bool pauseAtStart;
-        public bool persistLogs;
-        public bool activeActive;
-        public long logTriggerSizeMB;
-        public string storageConnectionString;
-        public long currentVersion;
-        public long upgradeToVersion;
-    }
-
-    public static class AmbrosiaRuntimeParms
-    {
-        public static bool _looseAttach = false;
-    }
-
-    public class AmbrosiaRuntime : VertexBase
+    public class AmbrosiaRuntime
     {
 #if _WINDOWS
         [DllImport("Kernel32.dll", CallingConvention = CallingConvention.Winapi)]
@@ -3849,35 +3827,6 @@ namespace Ambrosia
         {
         }
 
-        public override void Initialize(object param)
-        {
-            // Workaround because of parameter type limitation in CRA
-            AmbrosiaRuntimeParams p = new AmbrosiaRuntimeParams();
-            XmlSerializer xmlSerializer = new XmlSerializer(p.GetType());
-            using (StringReader textReader = new StringReader((string)param))
-            {
-                p = (AmbrosiaRuntimeParams)xmlSerializer.Deserialize(textReader);
-            }
-
-            bool sharded = false;
-
-            Initialize(
-                p.serviceReceiveFromPort,
-                p.serviceSendToPort,
-                p.serviceName,
-                p.serviceLogPath,
-                p.createService,
-                p.pauseAtStart,
-                p.persistLogs,
-                p.activeActive,
-                p.logTriggerSizeMB,
-                p.storageConnectionString,
-                p.currentVersion,
-                p.upgradeToVersion,
-                sharded
-            );
-        }
-
         internal void RuntimeChecksOnProcessStart()
         {
             if (!_createService)
@@ -3931,9 +3880,13 @@ namespace Ambrosia
                        string storageConnectionString,
                        long currentVersion,
                        long upgradeToVersion,
-                       bool sharded
+                       bool sharded,
+                       CRAClientLibrary coral,
+                       Action<string, IAsyncVertexInputEndpoint> addInput,
+                       Action<string, IAsyncVertexOutputEndpoint> addOutput
                        )
         {
+
             _runningRepro = false;
             _currentVersion = currentVersion;
             _upgradeToVersion = upgradeToVersion;
@@ -3956,7 +3909,7 @@ namespace Ambrosia
             _serviceName = serviceName;
             _storageConnectionString = storageConnectionString;
             _sharded = sharded;
-            _coral = ClientLibrary;
+            _coral = coral;
 
             Console.WriteLine("Logs directory: {0}", _serviceLogPath);
 
@@ -3971,10 +3924,11 @@ namespace Ambrosia
                     createService = true;
                 }
             }
-            AddAsyncInputEndpoint(AmbrosiaDataInputsName, new AmbrosiaInput(this, "data"));
-            AddAsyncInputEndpoint(AmbrosiaControlInputsName, new AmbrosiaInput(this, "control"));
-            AddAsyncOutputEndpoint(AmbrosiaDataOutputsName, new AmbrosiaOutput(this, "data"));
-            AddAsyncOutputEndpoint(AmbrosiaControlOutputsName, new AmbrosiaOutput(this, "control"));
+
+            addInput(AmbrosiaDataInputsName, new AmbrosiaInput(this, "data"));
+            addInput(AmbrosiaControlInputsName, new AmbrosiaInput(this, "control"));
+            addOutput(AmbrosiaDataOutputsName, new AmbrosiaOutput(this, "data"));
+            addOutput(AmbrosiaControlOutputsName, new AmbrosiaOutput(this, "control"));
             _createService = createService.Value;
             RecoverOrStartAsync().Wait();
         }
@@ -4060,7 +4014,7 @@ namespace Ambrosia
 
                     try
                     {
-                        if (client.DefineVertex(param.AmbrosiaBinariesLocation, () => new AmbrosiaRuntime()) != CRAErrorCode.Success)
+                        if (client.DefineVertex(param.AmbrosiaBinariesLocation, () => new AmbrosiaNonShardedRuntime()) != CRAErrorCode.Success)
                         {
                             throw new Exception();
                         }
